% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_matrix.R
\name{aux_matrix}
\alias{aux_matrix}
\title{Build our auxiliary matrices to estimate entry models}
\usage{
aux_matrix(data, y, max)
}
\arguments{
\item{data}{A \code{data.frame} object containing your data}

\item{y}{A string indicating the outcome variable}

\item{max}{An \code{integer} indicating the maximum number of competitors}
}
\value{
A list of the auxiliary matrices
}
\description{
Build our auxiliary matrices to estimate entry models
}
