% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br1.R
\name{br1}
\alias{br1}
\title{Build our optimization function}
\usage{
br1(params, n, N_max, l_params, A1, A2, S, N)
}
\arguments{
\item{params}{Parameters to construct function}

\item{n}{Number of observations in \code{data}}

\item{N_max}{An \code{integer} indicating the maximum number of competitors}

\item{l_params}{Length of parameters vector}

\item{A1}{Auxiliary matrix A1}

\item{A2}{Auxiliary matrix A2}

\item{S}{Size of the market}

\item{N}{Vector of zeros}
}
\value{
The function to be optimized
}
\description{
Build our optimization function
}
