% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br2.R
\name{br2}
\alias{br2}
\title{Build our optimization function}
\usage{
br2(params, n, N_max, A1, A2, S1, S2, N)
}
\arguments{
\item{params}{Parameters to construct function}

\item{n}{Number of observations in \code{data}}

\item{N_max}{An \code{integer} indicating the maximum number of competitors}

\item{A1}{Auxiliary matrix A1}

\item{A2}{Auxiliary matrix A2}

\item{S1}{First variable for size of the market}

\item{S2}{Second variable for size of the market}

\item{N}{Vector of zeros}
}
\value{
The function to be optimized
}
\description{
Build our optimization function
}
