% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_basic.R
\name{em_basic}
\alias{em_basic}
\title{Estimate our (basic) entry model with only one variable for the market size}
\usage{
em_basic(data, Sm, y, max = 5)
}
\arguments{
\item{data}{A \code{data.frame} object containing your data}

\item{Sm}{A string indicating the market size variable, present in \code{data}}

\item{y}{A string indicating the outcome variable, present in \code{data}}

\item{max}{An \code{integer} indicating the maximum number of competitors}
}
\value{
A tibble of the critical market sizes, as explained in Bresnahan and Reiss (1991)
}
\description{
Estimate our (basic) entry model with only one variable for the market size
}
