% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br1.R
\name{br1}
\alias{br1}
\title{Build our optimization function}
\usage{
br1(params, n, N_max, params0, A1, A2, S, N)
}
\arguments{
\item{params}{Parameters to construct function}

\item{n}{Number of observations}

\item{N_max}{An \code{integer} indicating the maximum number of competitors}

\item{params0}{Container for parameters}

\item{A1}{Auxiliary matrix A1}

\item{A2}{Auxiliary matrix A2}

\item{S}{Size of the market}

\item{N}{An \code{integer} indicating the competitor}
}
\value{
A list of the auxiliary matrices
}
\description{
Build our optimization function
}
