% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_2var.R
\name{em_2var}
\alias{em_2var}
\title{Two-Variable Entry Model}
\usage{
em_2var(data, Sm1, Sm2, y, N_max = 5)
}
\arguments{
\item{data}{A \code{data.frame} object containing your data}

\item{Sm1}{A string indicating the main market size variable, present in \code{data}}

\item{Sm2}{A string indicating the second market size variable, present in \code{data}}

\item{y}{A string indicating the outcome variable, present in \code{data}}

\item{N_max}{An \code{integer} indicating the maximum number of competitors. Defaults to 5.}
}
\value{
A tibble with critical market sizes and estimated parameters, as explained in Bresnahan and Reiss (1991)
}
\description{
Estimate entry model with two variables for the market size.
}
\examples{
tb <- data.frame(Sm1 = 1:5, Sm2 = 1:5, y = 1:5)

# estimate default model
em_n5 <- em_2var(tb, "Sm1", "Sm2", "y")

# estimate model with 3 competitors only
em_n3 <- em_2var(tb, "Sm", "Sm2", "y", N_max = 3)

\dontrun{
tb <- load_example_data()
em <- em_2var(tb, "Populacao", "RendaPerCapita", "n_agencias")
}

}
\references{
Bresnahan, T. F., & Reiss, P. C. (1991). Entry and competition in concentrated markets. Journal of political economy, 99(5), 977-1009.
}
\author{
Guilherme N. Jardim, Department of Economics, Pontifical Catholic University of Rio de Janeiro
}
