% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_basic.R
\name{em_basic}
\alias{em_basic}
\title{Basic Entry Model}
\usage{
em_basic(data, Sm, y, N_max = 5)
}
\arguments{
\item{data}{A \code{data.frame} object containing your data}

\item{Sm}{A string indicating the market size variable, present in \code{data}}

\item{y}{A string indicating the outcome variable, present in \code{data}}

\item{N_max}{An \code{integer} indicating the maximum number of competitors. Defaults to 5.}
}
\value{
A tibble of the critical market sizes, as explained in Bresnahan and Reiss (1991)
}
\description{
Estimate our (basic) entry model with only one variable for the market size
}
\examples{
tb <- load_example_data()
em <- em_basic(tb, "Populacao", "n_agencias")
}
